Weapon functionality moved into Limb
Weapon functionality appears to now be working the exact same as before
with the new weapon class and weapon.

Limb blueprint
Copied Weapon EventGraph to a new EventGraph in Limb called
WeaponEventGraph
New variables IsLimb and IsWeapon
Variable type for LimbWeaponEquipped changed to Limb
Input for SetLimbVariablesonWeaponEquip changed to Limb
Overlap on Limb changed to check if it's a a limb or weapon and call the
correct functions accordingly.

Skele-BP blueprint
Variable type for EquippedWeapon changed to Limb

NewWeapon (temporary)
Child of Limb

NewSword blueprint (temporary)
Child of NewWeapon

Both new classes newweapon and newsword will be deleted in my next
commit and the old ones will just be reparented and hopefully have all
the correct functionality working still.