Limbs working as weapons
Limbs can be picked up as weapons if there are no available sockets left
to be used as limbs. They act the exact same as weapons do.

Changes:

Limb blueprint
New function TryEquipAsWeapon - Called after each failed attempt to
attach
limb to player as a limb
Changes to OnRep_IsAttached

Issues:
Animations on the limbs keep playing when they are being used as
weapons.

The socket for the weapons all use SwordHand, which is not positioned
well for Limbs, so there needs to be a new socket created for limb
weapons.