Changed how Limbs of the same colour are counted
Changed where the limbs are counted and how they are counter. Much
tidier and makes a lot more sense. A lot more efficient too, not sure
why I did it the other way, no idea what I was thinking.

Changes:

Skele_BP blueprint
Made NumSameColour rep notify and added functionality to print if the
player has all matching limbs in there

Limb blueprint
Moved functionality for counting the matching limbs to the On Rep Is
Attached
function

Skele_GameMode
Removed the CheckNumberOfOwnColour event

Skele_FunctionLibrary
Added IsSameColour? function to check if the colour of the limb and
player inputs are the same.