New Bomb Spawn Points
Bombs now only explode when they have been thrown, instead of when they
are spawned. New BombSpawnPoint TargetPoint Actors which can be placed
around the map. 1 MUST be set to Initial Spawn for it to work.

Bombs will spawn in one place, and then randomly spawn in one of the
other spawns once the bombs are all taken.

Public variables for the spawn points include are the number of bombs
and the time between respawning.

I tried, for the last almost 2 or 3 hours to get a light working that
turns on and off when the spawn point is in use, but the damn
replication was not working at all for me, so I've given up for now.
Might be easiest to have a particle effect on the bombs when they are
just on the ground so they are easier to see.

Changes:

Skele_GameState
new variable "TempBombSpawnPoint"
2 new events to handle spawning the bombs

BombSpawnPoint
new blueprint target point

BombBP
new BombThrown variable
new BombSpawnPoint variable
Changed so it doesn't start ticking when spawned, waits until it's
thrown by the player
OnRepBombThrown calls the LightFuse Event and sets visibility of the
particle effects to true

Skele_BP
Added functionality in collision event to handle picking up bombs
Changes in ThrowBomb function