Shovel is working correctly
I made a new class for the shovel, so we still have the option of the
sword if we want at some point.

The issues are all fixed, I just had to add more to the construction
script since the settings on the side aren't working correctly.

Changes:
Sword_BP
Added in more of the settings to override the default limb ones in the
construction script

Shovel_BP blueprint
New blueprint created from limb. All same settings as sword_bp but it
has the different mesh

Skele_GameState blueprint
Changed it from spawning Sword_BP's to Shovel_BP's