Bomb Spawn Points replicating correctly
To use. Place down on the map, if you want it to be one of the initial
spawns, set Initial Spawn to true. If you only want 1 active at a time,
set only 1 to initial spawn. If you want 2, set 2 to initial spawn. And
so on. The Respawn Time is how long after that one has been used will
another one be opened. If none are selected for Initial spawn, none of
them will work, so make sure at least one is!