Updated Bomb coffin
Added Normal Map (looks great i rekon)
Added a mask for the bomb lid, so that if we wanted we could make it
flash or emit light before it is going to open or something like that.
or it could just emit light as well when it's open.
New coffin texture has an image of a bomb embedded/branded into it